+++ cfdm/test/test_functions.py
@@ -383,7 +383,16 @@ class FunctionsTest(unittest.TestCase):
         fields = [f, f, g]
         domains = [x.domain for x in (f, f, g)]
 
-        self.assertEqual(len(cfdm.unique_constructs(domains)), 2)
+        u = cfdm.unique_constructs(domains)
+        self.assertEqual(len(u), 2)
+        self.assertIsNot(u[0], domains[0])
+
+        u = cfdm.unique_constructs(domains, copy=False)
+        self.assertIs(u[0], domains[0])
+
+        u = cfdm.unique_constructs(domains, ignore_properties="z", copy=False)
+        self.assertIsNot(u[0], domains[0])
+
         self.assertEqual(len(cfdm.unique_constructs(domains + fields)), 4)
         self.assertEqual(
             len(cfdm.unique_constructs(domains + fields + [f.domain])), 4
@@ -391,13 +400,39 @@ class FunctionsTest(unittest.TestCase):
 
         # Test generator
         domains = (x.domain for x in ())
-        self.assertEqual(len(cfdm.unique_constructs(domains)), 0)
+        u = cfdm.unique_constructs(domains)
+        self.assertIsInstance(u, type(domains))
+        self.assertEqual(len(tuple(u)), 0)
 
         domains = (x.domain for x in (f,))
-        self.assertEqual(len(cfdm.unique_constructs(domains)), 1)
+        u = cfdm.unique_constructs(domains)
+        self.assertIsInstance(u, type(domains))
+        self.assertEqual(len(tuple(u)), 1)
 
         domains = (x.domain for x in (f, f, g))
-        self.assertEqual(len(cfdm.unique_constructs(domains)), 2)
+        u = cfdm.unique_constructs(domains)
+        self.assertIsInstance(u, type(domains))
+        self.assertEqual(len(tuple(u)), 2)
+
+        self.assertIsInstance(cfdm.unique_constructs(fields), type(fields))
+        self.assertIsInstance(cfdm.unique_constructs(tuple(fields)), tuple)
+
+        f.set_properties({"project": "model", "foo": "bar"})
+        f.set_property("project", "model")
+        f2 = f.copy()
+        f2.set_property("project", "obs")
+        g.set_properties({"project": "model", "foo": "bar"})
+        fields = [f, f2, g]
+
+        self.assertEqual(len(cfdm.unique_constructs(fields)), 3)
+        u = cfdm.unique_constructs(
+            fields, ignore_properties=["project", "foo"]
+        )
+        self.assertEqual(len(u), 2)
+        self.assertIsNone(u[0].get_property("project", None))
+        self.assertEqual(u[0].get_property("foo"), "bar")
+        self.assertEqual(u[1].get_property("project"), "model")
+        self.assertEqual(u[1].get_property("foo"), "bar")
 
     def test_context_managers(self):
         """Test the context manager support of the functions."""