+++ cfdm/data/netcdfarray.py
@@ -24,6 +24,7 @@ class NetCDFArray(FileArrayMixin, abstract.Array):
         mask=True,
         units=False,
         calendar=False,
+        missing_values=None,
         source=None,
         copy=True,
     ):
@@ -106,6 +107,13 @@ class NetCDFArray(FileArrayMixin, abstract.Array):
 
                 .. versionadded:: (cfdm) 1.10.0.1
 
+            missing_values: `dict`, optional
+                The missing value indicators defined by the netCDF
+                variable attributes. See `get_missing_values` for
+                details.
+
+                .. versionadded:: (cfdm) 1.10.0.3
+
             {{init source: optional}}
 
                 .. versionadded:: (cfdm) 1.10.0.0
@@ -153,6 +161,11 @@ class NetCDFArray(FileArrayMixin, abstract.Array):
             except AttributeError:
                 calendar = False
 
+            try:
+                missing_values = source._get_component("missing_values", None)
+            except AttributeError:
+                missing_values = None
+
         if shape is not None:
             self._set_component("shape", shape, copy=False)
 
@@ -167,6 +180,11 @@ class NetCDFArray(FileArrayMixin, abstract.Array):
                 address = (address,)
 
             self._set_component("address", address, copy=False)
+
+        if missing_values is not None:
+            self._set_component(
+                "missing_values", missing_values.copy(), copy=False
+            )
 
         self._set_component("dtype", dtype, copy=False)
         self._set_component("mask", mask, copy=False)
@@ -530,6 +548,41 @@ class NetCDFArray(FileArrayMixin, abstract.Array):
     #
     #        """
     #        return self._get_component("varid", None)
+
+    def get_missing_values(self):
+        """The missing value indicators from the netCDF variable.
+
+        .. versionadded:: (cfdm) 1.10.0.3
+
+        :Returns:
+
+            `dict` or `None`
+                The missing value indicators from the netCDF variable,
+                keyed by their netCDF attribute names. An empty
+                dictionary signifies that no missing values are given
+                in the file. `None` signifies that the missing values
+                have not been set.
+
+        **Examples**
+
+        >>> a.get_missing_values()
+        None
+
+        >>> b.get_missing_values()
+        {}
+
+        >>> c.get_missing_values()
+        {'missing_value': 1e20, 'valid_range': (-10, 20)}
+
+        >>> d.get_missing_values()
+        {'valid_min': -999}
+
+        """
+        out = self._get_component("missing_values", None)
+        if out is None:
+            return
+
+        return out.copy()
 
     def close(self, dataset):
         """Close the dataset containing the data.