+++ cfdm/test/test_NetCDFArray.py
@@ -1,5 +1,8 @@
+import atexit
 import datetime
 import faulthandler
+import os
+import tempfile
 import unittest
 
 faulthandler.enable()  # to debug seg faults and timeouts
@@ -7,8 +10,28 @@ faulthandler.enable()  # to debug seg faults and timeouts
 import cfdm
 
 
-class NetCDFArrayTest(unittest.TestCase):
-    """Unit test for the NetCDFArray class."""
+n_tmpfiles = 1
+tmpfiles = [
+    tempfile.mkstemp("_test_netCDF.nc", dir=os.getcwd())[1]
+    for i in range(n_tmpfiles)
+]
+(tmpfile,) = tmpfiles
+
+
+def _remove_tmpfiles():
+    """Remove temporary files created during tests."""
+    for f in tmpfiles:
+        try:
+            os.remove(f)
+        except OSError:
+            pass
+
+
+atexit.register(_remove_tmpfiles)
+
+
+class NetCDFTest(unittest.TestCase):
+    """Unit test for the NetCDF class."""
 
     def setUp(self):
         """Preparations called immediately before each test method."""
@@ -48,9 +71,34 @@ class NetCDFArrayTest(unittest.TestCase):
         a = cfdm.NetCDFArray()
         self.assertEqual(a.get_filenames(), ())
 
+    def test_NetCDFArray_get_missing_values(self):
+        """Test NetCDFArray.get_missing_values."""
+        f = cfdm.example_field(0)
+
+        f.set_property("missing_value", -999)
+        f.set_property("_FillValue", -3)
+        f.set_property("valid_range", [-111, 222])
+        cfdm.write(f, tmpfile)
+
+        g = cfdm.read(tmpfile)[0]
+        self.assertEqual(
+            g.data.source().get_missing_values(),
+            {
+                "missing_value": -999.0,
+                "_FillValue": -3,
+                "valid_range": (-111, 222),
+            },
+        )
+
+        c = g.coordinate("latitude")
+        self.assertEqual(c.data.source().get_missing_values(), {})
+
+        a = cfdm.NetCDFArray("file.nc", "ncvar")
+        self.assertIsNone(a.get_missing_values())
+
 
 if __name__ == "__main__":
     print("Run date:", datetime.datetime.now())
     cfdm.environment()
-    print()
+    print("")
     unittest.main(verbosity=2)