+++ Changelog.rst
@@ -1,3 +1,13 @@
+Version 1.10.?.?
+----------------
+
+**2023-??-??**
+
+* Fixed bug that caused `cf.Data.second_element` to fail for some data
+  shapes (https://github.com/NCAS-CMS/cfdm/issues/249)
+
+----
+
 Version 1.10.0.3
 ----------------